feat: replace `slowapi` rate limiter with `fastapi-limiter`, remove constants, move books api section to separate router and make programming books class a dependency
