+++ app/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "1.4.10"
+__version__ = "1.4.12"
